+++ src/constants/localStorageKeyNames.js
@@ -1,4 +1,4 @@
-const STORAGE_KEY_NAMES = {
+export const STORAGE_KEY_NAMES = {
   MARKED_USERS: 'whMarkedUsers',
   UNIQUE_USERS: 'whUniqueNicks',
   WH_SETTINGS: 'whSettings',